+++ fabric/src/main/java/io/github/sakurawald/fuji/module/mixin/world/persist/WorldGenSettingsMixin.java
@@ -1,12 +1,23 @@
 package io.github.sakurawald.fuji.module.mixin.world.persist;
 
+import com.google.common.collect.Maps;
+import io.github.sakurawald.fuji.module.initializer.world.accessor.ExtendedDimensionOptions;
+import io.github.sakurawald.fuji.module.initializer.world.structure.registry.FilteredRegistry;
+import java.util.Map;
+import net.minecraft.registry.Registry;
+import net.minecraft.registry.RegistryKey;
+import net.minecraft.world.dimension.DimensionOptions;
+import net.minecraft.world.dimension.DimensionOptionsRegistryHolder;
 import net.minecraft.world.level.WorldGenSettings;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.ModifyArg;
+import org.spongepowered.asm.mixin.injection.ModifyVariable;
 
 #if MC_VER <= MC_1_20_4
 #elif MC_VER > MC_1_20_4
 import com.google.common.collect.Maps;
-import io.github.sakurawald.fuji.module.initializer.world.accessor.IDimensionOptions;
+import io.github.sakurawald.fuji.module.initializer.world.accessor.ExtendedDimensionOptions;
 import io.github.sakurawald.fuji.module.initializer.world.structure.registry.FilteredRegistry;
 import net.minecraft.registry.Registry;
 import net.minecraft.registry.RegistryKey;
@@ -24,12 +35,28 @@ public class WorldGenSettingsMixin {
 
     #if MC_VER <= MC_1_20_4
     // FIXME add functions to filter the fuji dimensions
+
+    @ModifyVariable(method = "encode(Lcom/mojang/serialization/DynamicOps;Lnet/minecraft/world/gen/GeneratorOptions;Lnet/minecraft/world/dimension/DimensionOptionsRegistryHolder;)Lcom/mojang/serialization/DataResult;", at = @At("HEAD"), argsOnly = true)
+    private static DimensionOptionsRegistryHolder f(DimensionOptionsRegistryHolder dimensionOptionsRegistryHolder) {
+        Registry<DimensionOptions> dimensions = dimensionOptionsRegistryHolder.comp_1014();
+        FilteredRegistry<DimensionOptions> filteredDimensions = new FilteredRegistry<>(dimensions, ExtendedDimensionOptions.SAVE_PROPERTIES_PREDICATE);
+        return new DimensionOptionsRegistryHolder(filteredDimensions);
+    }
+
+    @SuppressWarnings("UnnecessaryLocalVariable")
+    @ModifyArg(method = "encode(Lcom/mojang/serialization/DynamicOps;Lnet/minecraft/world/gen/GeneratorOptions;Lnet/minecraft/registry/DynamicRegistryManager;)Lcom/mojang/serialization/DataResult;", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/dimension/DimensionOptionsRegistryHolder;<init>(Lnet/minecraft/registry/Registry;)V"), index = 0)
+    private static Registry<DimensionOptions> g(Registry<DimensionOptions> comp_1014) {
+        Registry<DimensionOptions> dimensions = comp_1014;
+        FilteredRegistry<DimensionOptions> filteredDimensions = new FilteredRegistry<>(dimensions, ExtendedDimensionOptions.SAVE_PROPERTIES_PREDICATE);
+        return filteredDimensions;
+    }
+
     #elif MC_VER > MC_1_20_4
     @ModifyArg(method = "encode(Lcom/mojang/serialization/DynamicOps;Lnet/minecraft/world/gen/GeneratorOptions;Lnet/minecraft/world/dimension/DimensionOptionsRegistryHolder;)Lcom/mojang/serialization/DataResult;"
         , at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/WorldGenSettings;<init>(Lnet/minecraft/world/gen/GeneratorOptions;Lnet/minecraft/world/dimension/DimensionOptionsRegistryHolder;)V"), index = 1)
     private static @NotNull DimensionOptionsRegistryHolder $wrapWorldGenSettings(DimensionOptionsRegistryHolder original) {
         Map<RegistryKey<DimensionOptions>, DimensionOptions> dimensions = original.comp_1014();
-        var saveDimensions = Maps.filterEntries(dimensions, entry -> IDimensionOptions.SAVE_PROPERTIES_PREDICATE.test(entry.getValue()));
+        var saveDimensions = Maps.filterEntries(dimensions, entry -> ExtendedDimensionOptions.SAVE_PROPERTIES_PREDICATE.test(entry.getValue()));
         return new DimensionOptionsRegistryHolder(saveDimensions);
     }
 
@@ -37,7 +64,7 @@ public class WorldGenSettingsMixin {
     @ModifyArg(method = "encode(Lcom/mojang/serialization/DynamicOps;Lnet/minecraft/world/gen/GeneratorOptions;Lnet/minecraft/registry/DynamicRegistryManager;)Lcom/mojang/serialization/DataResult;"
         , at = @At(value = "INVOKE", target = "Lnet/minecraft/world/dimension/DimensionOptionsRegistryHolder;<init>(Lnet/minecraft/registry/Registry;)V"), index = 0)
     private static Registry<DimensionOptions> $wrapWorldGenSettings(Registry<DimensionOptions> registry) {
-        return new FilteredRegistry<>(registry, IDimensionOptions.SAVE_PROPERTIES_PREDICATE);
+        return new FilteredRegistry<>(registry, ExtendedDimensionOptions.SAVE_PROPERTIES_PREDICATE);
     }
     #endif
 }